Adds ability to turn SSD1331 display on or off programmatically
When SSD1331 OLEDs are left on for long periods they can suffer from
burn-in. Whether this is common to all OLEDs or is peculiar to SSD1331s
(or even just a single batch of SSD1331s) is unknown. This change adds
an enableDisplay(boolean) function that permits a sketch to control
whether the display is lit or dormant. Using this new function, sketches
can turn the display off or on at predetermined intervals, or in
response to external signals such as button presses. Updates sent to a
dormant display are still processed normally and appear when the display
is next lit. A new test sketch demonstrates toggling the display on and
off in response to button presses.